requirements.txt
pandas
numpy
streamlit
openpyxl
python-dotenv
azure-storage-blob
azure-search-documents
langchain
langchain-openai
tiktoken